build(k82)!: provide migrations as a job
This translates much better to production deployment as these can be
applied at will instead of going through an "exec" route.

In this context this will now run automatically upon bootstrapping with
tilt and does not require a custom UI button as the job can be restarted
with already provided features.